# docs/openapi-external.yaml
openapi: 3.0.3
info:
  title: meWork360 Platform API — External v1
  version: 1.0.0-draft
  description: |
    **Contrato externo estável** do control plane meWork360 (`work-platform-api`).
    Modela **capabilities de negócio** (Tenant, Apps, Users, Jobs) — nunca subcomandos
    OCC, exit codes ou termos do protocolo Nextcloud interno.

    ## Status
    **DRAFT (ISSUE-038)** — este spec define o alvo do Sprint 0; **não está implementado**
    em produção. Fonte de verdade code-first para corrigir o drift DOC-001/ISSUE-021.

    ## Modelo assíncrono
    Operações de mutação retornam **HTTP 202** com envelope `{ data, meta }`, onde
    `meta.job_id` identifica o job na réplica local e `meta.status_url` aponta para
    `GET /v1/jobs/{id}`. Acompanhe o job até estado terminal (`success`, `failed`,
    `cancelled`).

    ## Janela de readiness
    Após provisionamento, o tenant pode permanecer em `provisioning` ou
    `provisioning_finishing`. Enquanto nessa janela, `POST /v1/tenants/{slug}/users`
    retorna **503** `tenant_not_ready` com header `Retry-After` (segundos).

    ## Autorização (ISSUE-037 — pré-requisito)
    Todas as rotas exigem Bearer token (`Authorization: Bearer <api_key>`).
    Cada operação declara o **scope** mínimo via extensão `x-required-scope`.
    A aplicação de scopes e o **binding tenant↔principal** depende do middleware
    `VerifyExternalPrincipal` (ISSUE-037 / SEC-V1-001) — gate duro antes de abrir
    a v1 a integradores externos.

...

  contact:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /tenants/{slug}/apps:
    parameters:
      - $ref: "#/components/parameters/TenantSlug"
    post:
      tags: [Apps]
      summary: Habilitar apps
      description: |
        Enfileira `apps:enable` com todos os app IDs consolidados em um único job
        (CSV upstream). Ordem do array `apps` é preservada.

        **Scope:** `apps:write`.
      operationId: enableApps
      x-required-scope: apps:write
      x-internal-mapping: LifecycleAsyncAction (cmd apps:enable)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnableAppsRequest"
            examples:
              enable:
                value:
                  apps: [files, calendar, contacts]
      responses:
        "202":
          description: Job enfileirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncAppsAccepted"
              examples:
                accepted:
                  value:
                    data:
                      apps: [files, calendar, contacts]
                    meta:
                      job_id: "b2c3d4e5-f6a7-8901-bcde-f12345678901"
                      status_url: "/v1/jobs/b2c3d4e5-f6a7-8901-bcde-f12345678901"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "404":
          $ref: "#/components/responses/TenantNotFound"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
        "503":
          $ref: "#/components/responses/ClusterUnreachable"
  /tenants/{slug}/users:
    parameters:
      - $ref: "#/components/parameters/TenantSlug"
    post:
      tags: [Users]
      summary: Criar usuário
      description: |
        Enfileira `users:create`. Campos extras (senha, e-mail, grupos) via payload
        interno — não expõe subcomandos OCC.

        Retorna **503** `tenant_not_ready` enquanto `status` ∈
        `{provisioning, provisioning_finishing}`.

        **Scope:** `users:write`.
      operationId: createUser
      x-required-scope: users:write
      x-internal-mapping: LifecycleAsyncAction (cmd users:create)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
            examples:
              admin:
                value:
                  username: admin.acme
                  password: "Str0ngP@ss1"
                  display_name: Admin Acme
                  email: admin@acme.com.br
                  groups: [admin]
      responses:
        "202":
          description: Job enfileirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncAccepted"
              examples:
                accepted:
                  value:
                    data:
                      username: admin.acme
                    meta:
                      job_id: "c3d4e5f6-a7b8-9012-cdef-123456789012"
                      status_url: "/v1/jobs/c3d4e5f6-a7b8-9012-cdef-123456789012"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "404":
          $ref: "#/components/responses/TenantNotFound"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
        "503":
          $ref: "#/components/responses/TenantNotReady"
  /tenants/{slug}/users/{username}:
    parameters:
      - $ref: "#/components/parameters/TenantSlug"
      - $ref: "#/components/parameters/Username"
    delete:
      tags: [Users]
      summary: Remover usuário
      description: |
        Enfileira `users:delete` para o username informado.

        **Scope:** `users:write`.
      operationId: deleteUser
      x-required-scope: users:write
      x-internal-mapping: LifecycleAsyncAction (cmd users:delete)
      responses:
        "202":
          description: Job enfileirado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AsyncAccepted"
              examples:
                accepted:
                  value:
                    data:
                      username: admin.acme
                    meta:
                      job_id: "d4e5f6a7-b8c9-0123-def0-234567890123"
                      status_url: "/v1/jobs/d4e5f6a7-b8c9-0123-def0-234567890123"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/ForbiddenScope"
        "404":
          $ref: "#/components/responses/TenantNotFound"
        "409":
          $ref: "#/components/responses/IdempotencyConflict"
        "422":
          $ref: "#/components/responses/ValidationFailed"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/UpstreamUnavailable"
        "503":
          $ref: "#/components/responses/ClusterUnreachable"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key emitida pelo painel admin (`/api-keys`). O token deve incluir os
        scopes exigidos pela operação (`x-required-scope`). Binding tenant↔principal
        é aplicado pelo middleware `VerifyExternalPrincipal` (ISSUE-037).
  parameters:
    TenantSlug:
      name: slug
      in: path
      required: true
      description: |
        Identificador do tenant. Pattern `^[a-z0-9-]+$`, 3–64 caracteres.
        Underscore e maiúsculas são rejeitados com 422.
      schema:
        type: string
        pattern: "^[a-z0-9-]+$"
        minLength: 3
        maxLength: 64
      example: acme-corp
    Username:
      name: username
      in: path
      required: true
      description: Username Nextcloud. Pattern `^[a-zA-Z0-9._-]+$`, máx. 64 caracteres.
      schema:
        type: string
        pattern: "^[a-zA-Z0-9._-]+$"
        maxLength: 64
      example: admin.acme
...
  schemas:
    Envelope:
      type: object
      required: [data]
      properties:
        data:
          description: Recurso ou objeto de domínio da operação.
          type: object
          nullable: true
        meta:
          $ref: "#/components/schemas/Meta"
    Meta:
      type: object
      required: [job_id, status_url]
      properties:
        job_id:
          type: string
          format: uuid
          description: Identificador do job assíncrono na réplica local.
        status_url:
          type: string
          description: Path relativo para polling do job (`GET /v1/jobs/{id}`).
          example: "/v1/jobs/7c9e6679-7425-40de-944b-e07fc1f90ae7"
    AsyncAccepted:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [meta]
          properties:
            data:
              type: object
              description: Payload mínimo da operação (pode ser `{}`).
...
    AsyncAppsAccepted:
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          required: [data, meta]
          properties:
            data:
              type: object
              required: [apps]
              properties:
                apps:
                  type: array
                  items:
                    type: string
...
    DomainError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Código estável de erro de domínio (nunca expõe subcmd/exit_code).
              enum:
                - validation_failed
                - unauthenticated
                - forbidden_scope
                - tenant_not_found
                - tenant_not_ready
                - idempotency_conflict
                - state_conflict
                - cluster_unreachable
                - upstream_unavailable
                - capability_not_available
                - rate_limited
                - not_implemented
            message:
              type: string
              description: Mensagem legível para integradores (sem stack trace).
            retry_after:
              type: integer
              description: Segundos sugeridos antes de retry (espelha header Retry-After quando aplicável).
              minimum: 1
            details:
              type: object
              description: Metadados adicionais seguros (ex. existing_job_id, diff, field errors).
              additionalProperties: true
...
    EnableAppsRequest:
      type: object
      required: [apps]
      properties:
        apps:
          type: array
          minItems: 1
          items:
            type: string
            pattern: "^[a-z0-9_]+$"
            maxLength: 100
    CreateUserRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          pattern: "^[a-zA-Z0-9._-]+$"
          maxLength: 64
        password:
          type: string
          minLength: 8
          description: Mín. 8 caracteres, letras e números.
        display_name:
          type: string
          maxLength: 255
        email:
          type: string
          format: email
          maxLength: 255
        quota:
          type: string
          description: "Ex.: `5 GB`, `none`, `default`, `unlimited`."
        groups:
          type: array
          items:
            type: string
            maxLength: 256
        subadmin_groups:
          type: array
          items:
            type: string
            maxLength: 256
...
  responses:
    Unauthenticated:
      description: Token ausente ou inválido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DomainError"
          examples:
            unauthenticated:
              value:
                error:
                  code: unauthenticated
                  message: Authentication required.
    ForbiddenScope:
      description: Token válido mas sem scope ou binding tenant↔principal
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DomainError"
          examples:
            forbidden:
              value:
                error:
                  code: forbidden_scope
                  message: Insufficient scope or tenant binding for this operation.
    TenantNotFound:
      description: Tenant não encontrado na réplica local
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DomainError"
          examples:
            notFound:
              value:
                error:
                  code: tenant_not_found
                  message: Tenant not found.
    TenantNotReady:
      description: Tenant ainda em provisionamento — operação de usuário bloqueada
      headers:
        Retry-After:
          description: Segundos antes de tentar novamente.
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DomainError"
          examples:
            notReady:
              value:
                error:
                  code: tenant_not_ready
                  message: Tenant is not ready for user operations.
                  retry_after: 60
                  details:
                    status: provisioning_finishing
    IdempotencyConflict:
      description: Requisição duplicada com payload diferente dentro da janela de 24h
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DomainError"
          examples:
            conflict:
              value:
                error:
                  code: idempotency_conflict
                  message: A job with different parameters already exists for this operation.
                  details:
                    existing_job_id: "7c9e6679-7425-40de-944b-e07fc1f90ae7"
...
    ClusterUnreachable:
      description: Cluster inativo ou inacessível
      headers:
        Retry-After:
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DomainError"
          examples:
            unreachable:
              value:
                error:
                  code: cluster_unreachable
                  message: Target cluster is unreachable.
                  retry_after: 60
    UpstreamUnavailable:
      description: Falha no upstream (SSH/Agent) — sem expor exit_code ou subcmd
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DomainError"
          examples:
            upstream:
              value:
                error:
                  code: upstream_unavailable
                  message: Upstream platform temporarily unavailable.
...
    RateLimited:
      description: Limite de taxa excedido (120 req/min por padrão)
      headers:
        Retry-After:
          schema:
            type: integer
            example: 60
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DomainError"
          examples:
            limited:
              value:
                error:
                  code: rate_limited
                  message: Too many requests.
                  retry_after: 60
    ValidationFailed:
      description: Falha de validação de entrada
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/DomainError"
          examples:
            validation:
              value:
                error:
                  code: validation_failed
                  message: The given data was invalid.
                  details:
                    slug: ["Slug já em uso."]